+++ utils/utils.py
@@ -10,9 +10,8 @@ import yaml
 
 # from stable_baselines3.common import logger
 from stable_baselines3 import A2C, CMAES, DDPG, DQN, PPO, SAC, TD3, TQC
+from stable_baselines3.common.buffers import NstepReplayBuffer  # noqa: F401
 from stable_baselines3.common.monitor import Monitor
-
-# from stable_baselines3.common.cmd_util import make_atari_env
 from stable_baselines3.common.utils import set_random_seed
 from stable_baselines3.common.vec_env import DummyVecEnv, SubprocVecEnv, VecEnv, VecFrameStack, VecNormalize